// src/api/inventory/models/EquipmentCreate.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Hospital Equipment & Service Management API
 * API for managing hospital equipment inventory, service requests, and location tracking.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface EquipmentCreate
 */
export interface EquipmentCreate {
    /**
     * 
     * @type {string}
     * @memberof EquipmentCreate
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof EquipmentCreate
     */
    type: string;
    /**
     * Unique inventory/asset number
     * @type {string}
     * @memberof EquipmentCreate
     */
    inventoryNumber: string;
    /**
     * 
     * @type {Date}
     * @memberof EquipmentCreate
     */
    warrantyExpiry?: Date;
    /**
     * Current location ID
     * @type {string}
     * @memberof EquipmentCreate
     */
    locationId: string;
    /**
     * 
     * @type {string}
     * @memberof EquipmentCreate
     */
    notes?: string;
}

/**
 * Check if a given object implements the EquipmentCreate interface.
 */
export function instanceOfEquipmentCreate(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "name" in value;
    isInstance = isInstance && "type" in value;
    isInstance = isInstance && "inventoryNumber" in value;
    isInstance = isInstance && "locationId" in value;

    return isInstance;
}

export function EquipmentCreateFromJSON(json: any): EquipmentCreate {
    return EquipmentCreateFromJSONTyped(json, false);
}

export function EquipmentCreateFromJSONTyped(json: any, ignoreDiscriminator: boolean): EquipmentCreate {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'name': json['name'],
        'type': json['type'],
        'inventoryNumber': json['inventoryNumber'],
        'warrantyExpiry': !exists(json, 'warrantyExpiry') ? undefined : (new Date(json['warrantyExpiry'])),
        'locationId': json['locationId'],
        'notes': !exists(json, 'notes') ? undefined : json['notes'],
    };
}

export function EquipmentCreateToJSON(value?: EquipmentCreate | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'name': value.name,
        'type': value.type,
        'inventoryNumber': value.inventoryNumber,
        'warrantyExpiry': value.warrantyExpiry === undefined ? undefined : (value.warrantyExpiry.toISOString().substr(0,10)),
        'locationId': value.locationId,
        'notes': value.notes,
    };
}